+++ src/constants/routes.js
@@ -1,0 +1,13 @@
+// App
+export const MAIN_STACK = "MainStack";
+export const FAVOURITES_STACK = "FavouritesStack";
+
+// Itchio
+export const ITCHIO_STACK = "ItchioStack";
+export const ITCHIO_AUTH_ROUTE = "ItchioAuthRoute";
+export const ITCHIO_STORE_ROUTE = "ItchioStoreRoute";
+export const ITCHIO_OWNED_ROUTE = "ItchioOwnedRoute";
+
+// Playdate
+export const PLAYDATE_STACK = "PlaydateStack";
+export const PLAYDATE_AUTH_ROUTE = "PlaydateAuthRoute";