+++ src/helper/itchio.js
@@ -1,9 +1,24 @@
 import DomSelector from "react-native-dom-parser";
+import {
+  fetchGameDownloads,
+  fetchItchioTaggedGames,
+  fetchOwnedGames,
+} from "../api/itchio";
+import {ATTRIBUTES, QUERY} from "../constants/itchio";
 
-export async function getPotentialPlaydateGameNames(page = 1) {
-  const response = await fetch(
-    `https://itch.io/games/tag-playdate?page=${page}&format=json`,
-  );
+export async function getGames(authorization) {
+  const response = await fetchOwnedGames(authorization);
+  return response.json();
+}
+
+export async function getGameDownloads(game, authorization) {
+  const {game_id, id} = game;
+  const response = await fetchGameDownloads(game_id, id, authorization);
+  return response.json();
+}
+
+export async function getPotentialPlaydateGameNames(page) {
+  const response = await fetchItchioTaggedGames(page);
   const {content, num_items} = await response.json();
 
   if (num_items === 0) {
@@ -11,12 +26,20 @@ export async function getPotentialPlaydateGameNames(page = 1) {
   }
 
   const dom = DomSelector(content);
-  const games = dom.getElementsByClassName("game_cell_data");
+  const games = dom.getElementsByClassName(QUERY.GAME_DATA_CLASS);
   const processedGames = [];
   for (let i = 0; i < games.length; i++) {
-    const titleElement = games[i].getElementsByClassName("title");
-    // We should filter by Playdate
-    processedGames.push(titleElement[0].children[0].text);
+    const gameId = games[i].attributes[[ATTRIBUTES.GAME_ID]];
+    const titleElement = games[i].getElementsByClassName(
+      QUERY.GAME_TITLE_CLASS,
+    );
+    const imgElement = games[i].getElementsByClassName(QUERY.GAME_IMAGE_CLASS);
+    let game = {
+      id: gameId,
+      title: titleElement[0].children[0].text,
+      img: imgElement[0].attributes[ATTRIBUTES.GAME_IMG] || "",
+    };
+    processedGames.push(game);
   }
   return processedGames;
 }
@@ -25,7 +48,7 @@ export async function getAllPotentialPlaydateGameNames() {
   const allNames = new Set();
 
   let loop = true;
-  let page = 1;
+  let page = 8;
 
   while (loop) {
     const names = await getPotentialPlaydateGameNames(page);
@@ -37,6 +60,5 @@ export async function getAllPotentialPlaydateGameNames() {
     });
     page++;
   }
-  console.log("NAMES", [...allNames]);
   return [...allNames];
 }