+++ src/navigation/stacks/itchioStack/ItchioStack.jsx
@@ -1,0 +1,14 @@
+import {createNativeStackNavigator} from "@react-navigation/native-stack";
+import React from "react";
+
+const Stack = createNativeStackNavigator();
+
+const ItchioStack = () => {
+  return (
+    <>
+      <Stack.Navigator></Stack.Navigator>
+    </>
+  );
+};
+
+export default ItchioStack;