+++ CHANGELOG.md
@@ -2,9 +2,41 @@
 
 All notable changes to this project will be documented in this file.
 
-## 0.0.1 - 2014-05-31
+## 0.2.1 - 2022-09-30
+
+### Changed
+
+- Moved stacks to main stack
 
 ### Added
+
+- Added logic for authenticated routes for itchio
+
+## 0.2.0 - 2022-09-30
+
+### Added
+
+- Added React Navigation
+- Added Basic tab navigation
+- Added BaseScreen component for new screens
+
+## 0.1.0 - 2022-09-24
+
+### Added
+
+- Added dotenv and env example
+- Added OAuth for itchio
+- Added a basic list/grid view for some games (WIP)
+- Added basic login forms
+
+### Fixed
+
+- Fixed changelog dates
+
+## 0.0.1 - 2022-09-24
+
+### Added
+
 - First commit
-- This CHANGELOG file 
+- This CHANGELOG file
 - README file